OnnixBot with Payments/requirements_py313.txt
# Core requirements for Python 3.13
python-telegram-bot>=21.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
ta>=0.10.2
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
matplotlib>=3.8.0
seaborn>=0.13.0
schedule>=1.2.0
apscheduler>=3.10.4

# Skip problematic packages for now
# pillow  # Skip due to Python 3.13 issues
# plotly  # Skip due to Python 3.13 issues
# websocket-client  # Skip if not needed